+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Some dependencies that were used but not declared (often transient).
 
+### Changed
+
+- Replaced deprecated `multihashes` with `multiformats`.
+
 ### Removed
 
 - No more `assert` usage.